backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.12.0
httpx==0.27.2
python-dotenv==1.0.0
